+++ src/style/dump/dump_grid_gz.h
@@ -53,7 +53,7 @@ class DumpGridGz : public Dump {
 			      "rigid", "T",
 			      "ntypex", "ntypey", "ntypez"};
 private:
-  deque<pair<thread, vector<double>>>  threads;        ///< Pair storing the threads and the buffer
+  deque<pair<thread, vector<float>>>  threads;        ///< Pair storing the threads and the buffer
 
   void write_to_file(bigint, string, bigint, bigint);
 };
